+++ muna/beta/__init__.py
@@ -3,5 +3,6 @@
 #   Copyright © 2026 NatML Inc. All Rights Reserved.
 #
 
+from .annotations import Annotations
 from .metadata import *
 from .remote import RemoteAcceleration